refactor: introduce `add_services` module
The only code now that is specific to each service is creating them. There is potential here to also
make this process generic and move it into `ServiceManager`, but this branch is already big enough
as it is. For now, the service creation code is moving into the `add_services` module. This removes
the need for having separate `node_control`, `faucet_control` and `daemon_control` modules. The test
code for the `add` functions is quite large, so I've retained the use of a `tests.rs` file.
