+++ sn_node_manager/src/bin/daemon/main.rs
@@ -12,7 +12,7 @@ extern crate tracing;
 use clap::Parser;
 use color_eyre::{self, eyre::Result};
 use libp2p_identity::PeerId;
-use sn_node_manager::{config::get_node_registry_path, daemon_control::DAEMON_DEFAULT_PORT, rpc};
+use sn_node_manager::{config::get_node_registry_path, rpc, DAEMON_DEFAULT_PORT};
 use sn_service_management::{
     safenode_manager_proto::{
         safe_node_manager_server::{SafeNodeManager, SafeNodeManagerServer},
