+++ sn_node_manager/tests/daemon.rs
@@ -11,7 +11,7 @@ mod utils;
 use assert_cmd::Command;
 use color_eyre::eyre::{bail, OptionExt, Result};
 use libp2p::PeerId;
-use sn_node_manager::daemon_control::DAEMON_DEFAULT_PORT;
+use sn_node_manager::DAEMON_DEFAULT_PORT;
 use sn_service_management::safenode_manager_proto::{
     safe_node_manager_client::SafeNodeManagerClient, NodeServiceRestartRequest,
 };
